+++ tests/test_django.py
@@ -7,7 +7,7 @@ run::
 
 for a single test, run::
 
-   pytest tests/test_django.py::test_multiple_endblocks --cov=src/djlint \
+   pytest tests/test_django.py::test_complex_attributes --cov=src/djlint \
      --cov-branch --cov-report xml:coverage.xml --cov-report term-missing
 
 """
@@ -290,9 +290,14 @@ def test_complex_attributes(runner: CliRunner, tmp_file: TextIO) -> None:
         b"""<img data-src="{% if report.imgs.exists %}{{ report.imgs.first.get_absolute_url|size:"96x96"}}{% else %}{% static '/img/report_thumb_placeholder_400x300.png' %}{% endif %}" src="{% static '/img/loader.gif' %}" alt="report image"/>""",
     )
     assert output["exit_code"] == 1
+
     assert (
         output["text"]
-        == r"""<img data-src="{% if report.imgs.exists %}{{ report.imgs.first.get_absolute_url|size:"96x96" }}{% else %}{% static '/img/report_thumb_placeholder_400x300.png' %}{% endif %}"
+        == r"""<img data-src="{% if report.imgs.exists %}
+                   {{ report.imgs.first.get_absolute_url|size:"96x96" }}
+               {% else %}
+                   {% static '/img/report_thumb_placeholder_400x300.png' %}
+               {% endif %}"
      src="{% static '/img/loader.gif' %}"
      alt="report image"/>
 """
@@ -304,10 +309,22 @@ def test_complex_attributes(runner: CliRunner, tmp_file: TextIO) -> None:
         b"""<a class=" {% if favorite == "yes" %}favorite{% endif %} has-tooltip-arrow has-tooltip-right" data-tooltip=" {% if favorite == "yes" %}Remove from Favorites {% else %}Add to Favorites{% endif %}" fav-type="report" object-id="{{ report.report_id }}"><span class="icon has-text-grey is-large "><i class="fas fa-lg fa-star"></i></span></a>""",
     )
     assert output["exit_code"] == 1
+
     assert (
         output["text"]
-        == r"""<a class=" {% if favorite == "yes" %}favorite{% endif %} has-tooltip-arrow has-tooltip-right"
-   data-tooltip=" {% if favorite == "yes" %}Remove from Favorites {% else %}Add to Favorites{% endif %}"
+        == r"""<a class=" """
+        + """
+          {% if favorite == "yes" %}
+              favorite
+          {% endif %}
+          has-tooltip-arrow has-tooltip-right"
+   data-tooltip=" """
+        + """
+                 {% if favorite == "yes" %}
+                     Remove from Favorites
+                 {% else %}
+                     Add to Favorites
+                 {% endif %}"
    fav-type="report"
    object-id="{{ report.report_id }}">
     <span class="icon has-text-grey is-large ">
@@ -322,10 +339,15 @@ def test_complex_attributes(runner: CliRunner, tmp_file: TextIO) -> None:
         b"""<div class="media-content" {% ifchanged comment.stream_id %} comments-msg {% else %} comments-newMsgReply {% endifchanged %}>""",
     )
     assert output["exit_code"] == 1
+
     assert (
         output["text"]
         == r"""<div class="media-content"
-     {% ifchanged comment.stream_id %} comments-msg {% else %} comments-newMsgReply {% endifchanged %}>
+     {% ifchanged comment.stream_id %}
+         comments-msg
+     {% else %}
+         comments-newMsgReply
+     {% endifchanged %}>
 """
     )
     output = reformat(
@@ -337,6 +359,35 @@ def test_complex_attributes(runner: CliRunner, tmp_file: TextIO) -> None:
         output["text"]
         == """<a class="piwik_download"
    href="{% static activity_version.get_win_document_with_images_file_path %}?{% now "jSFYHi" %}">
+"""
+    )
+    assert output["exit_code"] == 1
+
+    output = reformat(
+        tmp_file,
+        runner,
+        b"""<span {%if a%}required{%endif%}title="{% if eev.status == eev.STATUS_CURRENT %} {% trans 'A' %} {% elif eev.status == eev.STATUS_APPROVED %} {% trans 'B' %} {% elif eev.status == eev.STATUS_EXPIRED %} {% trans 'C' %}{% endif %}" class="asdf {%if a%}b{%endif%} asdf" {%if a%}checked{%endif%}>""",
+    )
+    assert (
+        output["text"]
+        == """<span {% if a %}
+          required
+      {% endif %}
+      title="{% if eev.status == eev.STATUS_CURRENT %}
+                 {% trans 'A' %}
+             {% elif eev.status == eev.STATUS_APPROVED %}
+                 {% trans 'B' %}
+             {% elif eev.status == eev.STATUS_EXPIRED %}
+                 {% trans 'C' %}
+             {% endif %}"
+      class="asdf
+             {% if a %}
+                 b
+             {% endif %}
+             asdf"
+      {% if a %}
+          checked
+      {% endif %}>
 """
     )
     assert output["exit_code"] == 1